Minor update to the core library.

+++ KioskApp/MenuViewController_Kiosk.m
@@ -162,7 +162,7 @@
     
 	BookItemView * bookItemView = nil;
     
-	int documentsCount; // Used to iterate over each item in the list.
+	NSUInteger documentsCount; // Used to iterate over each item in the list.
 	
 	//Graphics visualization
